requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-hitcount==1.3.1
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
django-filter==2.2.0
djangorestframework==3.10.3
reportlab==3.5.26
sqlparse==0.3.0
urllib3==1.25.6